refactor: Rename useImageLoad_ for internal use
