+++ packages/insight-viewer/src/hooks/useImageLoad_/types.ts
@@ -1,3 +1,7 @@
 import { LOADING_STATE } from './const'
 
 export type LoadingState = typeof LOADING_STATE[keyof typeof LOADING_STATE]
+
+export interface ImageLoadState {
+  loadingState: LoadingState
+}
